chore: move coordinator v2 to coordinator
